+++ changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- New validator for autplaying videos (shortcut: auvid)
 - New validator for meta description (shortcut: mdesc)
 
 ## [0.3]